Modified schema modification to be inside of yamlSchemaService
