+++ src/languageservice/utils/schemaUrls.ts
@@ -1,0 +1,2 @@
+export const KUBERNETES_SCHEMA_URL = 'https://raw.githubusercontent.com/instrumenta/kubernetes-json-schema/master/v1.14.0-standalone-strict/all.json';
+export const JSON_SCHEMASTORE_URL = 'http://schemastore.org/api/json/catalog.json';
